YT_POST_GEN/requirements.txt
langchain_community==0.3.21
langchain_core==0.3.51
langchain_groq==0.3.2
langchain_ollama==0.3.1
langgraph==0.3.25
python-dotenv==1.1.0
streamlit==1.44.1
typing_extensions==4.13.1
youtube-transcript-api==1.0.3
uvicorn
